client/requirements.txt
python-socketio
pyserial
pynmea2